Moves logging function into settings file
